[FIX] mail: Avoid traceback when calling 'message_change_thread'
Purpose
=======

The method 'message_change_thread' transfers the list of the mail
thread messages from an model to another.

So as the mail.tracking.values. But it a mail.tracking.value is
linked to the field 'phone' for example, that doesn't exist on the
other model, this will crash when trying to compute the fields
groups.

In that case obfuscate the mail.tracking.value with the group_sytem,
so that potential group-protected values ('wage' for example) don't
show up on the chatter, but is still available for the super user.